+++ webgraph/src/graphs/bvgraph/comp/bvcomp.rs
@@ -184,8 +184,6 @@ impl Compressor {
             }
         }
         // write the residuals
-        // first signal the number of residuals to the encoder
-        writer.num_of_residuals(self.residuals.len());
         if !self.residuals.is_empty() {
             written_bits += writer
                 .write_first_residual((self.residuals[0] as i64 - curr_node as i64).to_nat())?