+++ webgraph/src/utils/mod.rs
@@ -132,10 +132,6 @@ impl<D: Decode, E: Encode> Decode for Converter<D, E> {
         self.offset += self.encoder.write_residual(res).unwrap();
         res
     }
-    #[inline(always)]
-    fn num_of_residuals(&mut self, num_of_residuals: usize) {
-        self.encoder.num_of_residuals(num_of_residuals);
-    }
 }
 
 /// An enum expressing the memory requirements for batched algorithms